Fix type and shape issue for the reference implementation (#4839)
* fix elu output type


* fix types and shapes for many operators


* fix other potential type issues


* remove comments


* lint


---------

+++ onnx/reference/ops/op_round.py
@@ -8,4 +8,4 @@ from ._op import OpRunUnaryNum
 
 class Round(OpRunUnaryNum):
     def _run(self, x):  # type: ignore
-        return (np.round(x),)
+        return (np.round(x).astype(x.dtype),)
